backend/requirements.txt
fastapi
uvicorn
python-socketio
langchain
langchain-community
langchain-openai
google-generativeai
chromadb
python-multipart
requests
